--- apps/example/src/app/layout.tsx
@@ -1,21 +1,0 @@
-import type { Metadata } from "next";
-
-import "@picocss/pico";
-import "./globals.css";
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
-}
