--- apps/example/src/app/page.tsx
@@ -1,5 +1,0 @@
-import { DefaultSampleForm } from "@/forms";
-
-export default function Home() {
-  return <DefaultSampleForm />;
-}
